Fix #15 and bump up version to 0.2.4
Creating large example projects with Naggy installed was causing the entire IDE to freeze. Tracked the problem down to a deadlock between Naggy's delayed background processing thread and the IDE GUI thread. The VS Shell provides a GetOrCreateSingletonProperty method with each buffer, and Naggy uses this to create and cache the ClangAdapter for that buffer. GetOrCreateSingletonProperty is protected by a lock, and is used by the IDE GUI thread too. The deadlock occurs when Naggy's background thread calls GetOrCreateSingletonProperty on a new buffer, (which uses information from DTE methods to create a new ClangAdapter for that buffer) and the UI thread at that moment calls GetOrCreateSingletonProperty. Because DTE is run in an STA, the call from Naggy returns only if the UI thread is available, but the UI thread is blocked on GetOrCreateSingletonProperty.
The fix was to use ThreadHelper.Invoke to marshal Naggy's GetOrCreateSingletonProperty calls to the STA thread.