Fix #8 - Naggy not working with AVR Studio 5.1
Overwrote previous commit - that code was committed when working with an upgraded 5.1 build, which bought in its own set of problems.
The actual problem is the presence of an extra bin directory in the list of default include paths reported by the toolchain. Fixed it by replacing bin\ with the empty string.