Fix issue #26
For a document that is not part of a project, try checking if it is from the standard toolchain path, and if yes, return the first project with that toolchain in the solution as the project for that document. Without this, the default include paths do not get populated for toolchain header files, and therefore the missing header diagnostics.