Fix issue #43
Clang internally was holding files open unless the CompilerInstance was destroyed. The destruction was being done only at the start of Process, which was resulting in some files becoming unsaveable.

The fix gathers all the necessary information from Clang within the Process method and destroys the CompilerInstance at the end.